+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the "Auctioneer Software" extension will be documented in
 
 ## [Unreleased](https://github.com/jacobSND/as-vscode/compare/v1.5.2...HEAD)
 
+### Added
+ - ⚙️ Settings: Repo Clone method (SSH, HTTPS, CLI)
+
 ## [1.5.2](https://github.com/jacobSND/as-vscode/compare/v1.4.4...v1.5.0) - 2025-09-13
 
 ### Fixed